write script that will create the ccs project
I think using the script will be easier than creating projects using
the CCS gui. The script gives me more control over the directory
structure of the repository and the project.

I considered making the project be the root-level of the repository, but
it is easier to create the project as a subfolder in the repository.
Using a subfolder makes it so the docs folder is not inside the CCS
project directory, but I think that is okay. We'll see what makes the
most sense as we move forward with more complicated projects.